requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
scikit-learn>=1.0.2
matplotlib>=3.5.1
numpy>=1.22.0
pillow>=9.0.0
plotly>=5.10.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0
seaborn>=0.12.0
xlrd>=2.0.1
odfpy>=1.4.1
pyxlsb>=1.0.9
waitress>=2.1.2
google-genai